requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
